// src/app/models/waterfall-ticker.ts
export class WaterFall {
    exchange: String = '';
    market: String = '';
    symbol: String = '';
    bidSize: Number = 0;
    bid: Number = 0;
    ask: Number = 0;
    askSize: Number = 0;
    flag: Boolean = false;
    lastVolume: Number = 0;
    turnover: Number = 0;
    trades: Number = 0;
    average: Number = 0;
    open: Number = 0;
    high: Number = 0;
    low: Number = 0;

    constructor() {

    }
}